Rewind: Explore performing description formatting as enhancement
This creates a formatted description for a white-listed set of events in
the API mapping in the data layer. It's a quick-and-dirtry way of adding
progressive enhancement to the app without needing to change the API
response.

Currently these are not being translated but we could easily add that
information. This also implies more code to maintain across the apps and
almost certainly should move into the API quickly. Until that time,
however, it's a reasonable first-approximation we can use to figure out
_what_ formatting is essential and how it should work.